assignment3/pyproject.toml
[pip]
requires = [
    "setuptools",
    # 4110 only:
    "cython",
    "numpy",
    "numba",
    "pillow",
    "line-profiler"
]
build-backend = "setuptools.build_meta"

[project]
version = "0.1.0"
requires-python = ">=3.7"
license = {text = "MIT License"}
name = "instapy"
description = "Helps to turn an image into grayscale or sepia modes."
readme = "README.md"
dependencies = [
    "cython",
    "numpy",
    "numba",
    "pillow",
    "line-profiler"
 ]


[project.scripts]
instapy = "instapy.cli:main"
